Example_service/src/requirements.txt
requests>=2.27.1
aniso8601==9.0.1
bosdyn-api==3.2.1.post1
bosdyn-client==3.2.1.post1
bosdyn-core==3.2.1.post1
bosdyn-mission==3.2.1.post1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
Flask==2.2.2
Flask-RESTful==0.3.9
grpcio==1.51.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
protobuf==4.21.9
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0